UXUI-195 addressing PR comments added how to add a ui component to the toolkit in README.md stop running browserify globally rename index.js to ui-toolkit.js
