+++ README.md
@@ -86,12 +86,41 @@ var UIToolkit = require('ui-toolkit');
 If you would like to cherry pick which components to include in your project, you can use this method:
 
 ```jsx
-var Button = require('ui-toolkit/ui-component-button');
+var Button = require('ui-toolkit').Button;
 <Button />
 
-var BootstrapPanel = require('ui-toolkit/react-bootstrap/lib/Panel');
+var BootstrapPanel = require('ui-toolkit').BootstrapPanel;
 <BootstrapPanel />
 ```
+
+Adding a new component
+---
+#### Adding a component to ui-toolkit
+1. Firstly (if you haven't already), create a new component using our [yeoman generator](https://github.com/holidayextras/generator-ui-component)
+2. Install your new component e.g. `npm install --save ui-component-foo`
+3. Open `./src/ui-toolkit.js`
+4. Add `UIToolkit.YourComponent = require('ui-component-foo');` to the `Custom Components` section
+
+#### Updating the docs (todo, simplify this step)
+1. In `./docs/examples/`, create a new jsx file that describes your component. e.g.
+```jsx
+// docs/examples/MyComponent.jsx
+var myComponentInstance = (
+  <UIToolkit.MyComponent foo="bar" />
+);
+React.render(myComponentInstance, mountNode);
+```
+
+2. In `./docs/src/Components.jsx`, add a new `<article>`, similar to the existing ones e.g.
+```jsx
+<article>
+  <h3 id="my-component">My component</h3>
+  <p>My component description</p>
+  <CustomComponent codeText={fs.readFileSync(__dirname + '/../examples/MyComponent.jsx', 'utf8')} />
+</article>
+```
+
+3. You should now be able to see your component when running `grunt docs`
 
 UI Toolkit Themes
 ---
@@ -120,7 +149,6 @@ You will only need to upload the following static assets once generated:
 + assets/
 + vendor/
 + index.html
-
 
 Grunt Terminal Commands:
 ---
