+++ docs/scripts/build.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 echo "Building docs"
-browserify -g [ reactify --extension '.jsx' ] ./client.js -o ./assets/bundle.js
+browserify -t [ reactify --extension '.jsx' ] ./client.js -o ./assets/bundle.js
